feat: add home route
